pyproject.toml
[tool.pdm]

[project]
name = "MeTTS Dataset"
version = "0.1.0"
description = "A dataset for TTS using measure values and priors."
authors = [
    {name = "Christoph Minixhofer", email = "christoph.minixhofer@gmail.com"},
]
dependencies = [
    "alignments>=0.0.9",
    "datasets>=2.8.0",
    "numpy>=1.23.5",
    "tqdm>=4.64.1",
    "pandas==1.3.0",
    "phones>=0.0.6",
    "torchaudio>=0.13.1",
    "scikit-learn>=1.2.0",
    "soundfile>=0.11.0",
    "librosa>=0.9.2",
    "transformers>=4.25.1",
]
requires-python = ">=3.8"
readme = "README.md"
license = {text = "MIT"}
[project.optional-dependencies]
measures = [
    "pyworld>=0.3.2",
    "SRMRpy @ git+https://github.com/MiniXC/SRMRpy.git",
    "textgrid>=1.5",
    "nnAudio>=0.3.2",
]

[build-system]
requires = ["pdm-pep517>=1.0"]
build-backend = "pdm.pep517.api"
